Modulariza configuracion y utilidades
