+++ js/main.js
@@ -1,7 +1,8 @@
 import { auth, db, doc, setDoc, getDoc, onSnapshot, updateDoc, serverTimestamp, onAuthStateChanged } from './firebase.js';
 import { showNotification, fitTextToContainer } from './utils.js';
-import { registerAuthHandlers, handleLogout } from './auth.js';
-import { registerNavigationEvents } from './navigation.js';
+import { userSettings, updateLocalSettings, populateSettingsForm, registerSettingsHandlers } from './settings.js';
+import { initCalculator } from './calculator.js';
+import { initRappicargo } from './rappicargo.js';
 
 const { jsPDF } = window.jspdf;
 
@@ -12,15 +13,6 @@ let scrollInterval = null;
 let scrollPauseTimeout = null;
 let oldOrdersInterval = null;
 let previousCodesCount = 0;
-let userSettings = {
-  blinkMinutes: 5,
-  criticalMinutes: 15,
-  scrollSpeed: 3,
-  viewerSize: 3,
-  calculatorSize: 4,
-  viewerFooterText: "⬅️ para retirar",
-  viewerFooterSize: 3,
-};
 
 const loginView = document.getElementById('login-view');
 const roleSelectionView = document.getElementById('role-selection-view');
@@ -102,7 +94,7 @@ function listenForCodes(userId) {
       sessionData = docSnap.data();
       sessionData.codes = sessionData.codes || [];
       sessionData.history = sessionData.history || [];
-      userSettings = { ...userSettings, ...sessionData.settings };
+      updateLocalSettings(sessionData.settings);
 
       const isNewCode = sessionData.codes.length > previousCodesCount;
       previousCodesCount = sessionData.codes.length;
@@ -456,11 +448,6 @@ if (closeGroupsEl) {
 }
 
 function backToMenu() { localStorage.removeItem('userRole'); showView(roleSelectionView); }
-document.getElementById('op-back-to-menu-btn').addEventListener('click', backToMenu);
-document.getElementById('viewer-back-to-menu-btn').addEventListener('click', backToMenu);
-document.getElementById('history-back-to-menu-btn').addEventListener('click', backToMenu);
-document.getElementById('settings-back-to-menu-btn').addEventListener('click', backToMenu);
-document.getElementById('closeorders-back-to-menu-btn').addEventListener('click', backToMenu);
 
 document.querySelectorAll('.keypad-btn').forEach(btn => btn.addEventListener('click', () => {
   const display = document.getElementById('code-display');
@@ -515,124 +502,9 @@ document.getElementById('delete-history-btn').addEventListener('click', async ()
     showNotification("Historial borrado correctamente.");
   }
 });
-
-function populateSettingsForm() {
-  document.getElementById('blink-minutes-input').value = userSettings.blinkMinutes;
-  document.getElementById('critical-minutes-input').value = userSettings.criticalMinutes;
-  document.getElementById('scroll-speed-input').value = userSettings.scrollSpeed;
-  document.getElementById('viewer-size-input').value = userSettings.viewerSize;
-  document.getElementById('calculator-size-input').value = userSettings.calculatorSize;
-  document.getElementById('viewer-text-input').value = userSettings.viewerFooterText;
-  document.getElementById('viewer-footer-size-input').value = userSettings.viewerFooterSize;
-}
-document.getElementById('save-settings-btn').addEventListener('click', async () => {
-  const userId = auth.currentUser?.uid;
-  if (!userId) return;
-  const newSettings = {
-    blinkMinutes: parseInt(document.getElementById('blink-minutes-input').value) || 5,
-    criticalMinutes: parseInt(document.getElementById('critical-minutes-input').value) || 15,
-    scrollSpeed: parseInt(document.getElementById('scroll-speed-input').value) || 3,
-    viewerSize: parseInt(document.getElementById('viewer-size-input').value) || 3,
-    calculatorSize: parseInt(document.getElementById('calculator-size-input').value) || 4,
-    viewerFooterText: document.getElementById('viewer-text-input').value || "⬅️ para retirar",
-    viewerFooterSize: parseInt(document.getElementById('viewer-footer-size-input').value) || 3,
-  };
-  const sessionDocRef = doc(db, "sessions", userId);
-  await updateDoc(sessionDocRef, { settings: newSettings });
-  showNotification("Configuración guardada.");
-  backToMenu();
-});
-
-const calculatorOverlay = document.getElementById('calculator-overlay');
-const calculatorModal = document.getElementById('calculator-modal');
-let calcState = { displayValue: '0', firstOperand: null, waitingForSecondOperand: false, operator: null, expression: '' };
-
-function updateCalcDisplay() {
-  document.getElementById('calc-display').textContent = calcState.displayValue;
-  document.getElementById('calc-expression-display').textContent = calcState.expression;
-}
-
-document.getElementById('open-calculator-btn').addEventListener('click', () => {
-  const calcSizes = ['w-64','w-72','w-80','w-96','w-[28rem]','w-[32rem]','w-[36rem]','w-[40rem]'];
-  calculatorModal.className = 'bg-white dark:bg-gray-800 p-4 rounded-lg shadow-2xl transition-transform transform open';
-  calculatorModal.classList.add(calcSizes[userSettings.calculatorSize - 1] || calcSizes[3]);
-  calculatorOverlay.classList.remove('hidden');
-});
-calculatorOverlay.addEventListener('click', (e) => {
-  if (e.target === calculatorOverlay) {
-    calculatorModal.classList.remove('open');
-    calculatorOverlay.classList.add('hidden');
-  }
-});
-document.getElementById('calc-buttons').addEventListener('click', (e) => {
-  const { target } = e;
-  if (!target.matches('button')) return;
-  const key = target.textContent;
-
-  if (key === 'C') {
-    calcState = { displayValue:'0', firstOperand:null, waitingForSecondOperand:false, operator:null, expression:'' };
-  } else if (!isNaN(parseFloat(key)) || key === '.') {
-    if (calcState.waitingForSecondOperand) {
-      calcState.displayValue = key;
-      calcState.waitingForSecondOperand = false;
-    } else {
-      calcState.displayValue = calcState.displayValue === '0' ? key : (calcState.displayValue + key);
-    }
-  } else if (['+','-','*','/'].includes(key)) {
-    const inputValue = parseFloat(calcState.displayValue);
-    if (calcState.operator && calcState.waitingForSecondOperand)  {
-      calcState.operator = key;
-      calcState.expression = `${calcState.firstOperand} ${key}`;
-      return;
-    }
-    if (calcState.firstOperand == null) {
-      calcState.firstOperand = inputValue;
-    } else if (calcState.operator) {
-      const result = performCalculation[calcState.operator](calcState.firstOperand, inputValue);
-      calcState.displayValue = `${parseFloat(result.toFixed(7))}`;
-      calcState.firstOperand = result;
-    }
-    calcState.waitingForSecondOperand = true;
-    calcState.operator = key;
-    calcState.expression = `${calcState.firstOperand} ${key}`;
-  } else if (key === '=') {
-    if (calcState.operator == null || calcState.waitingForSecondOperand) return;
-    const inputValue = parseFloat(calcState.displayValue);
-    calcState.expression = `${calcState.firstOperand} ${calcState.operator} ${inputValue} =`;
-    const result = performCalculation[calcState.operator](calcState.firstOperand, inputValue);
-    calcState.displayValue = `${parseFloat(result.toFixed(7))}`;
-    calcState.firstOperand = null;
-    calcState.operator = null;
-    calcState.waitingForSecondOperand = true;
-  }
-  updateCalcDisplay();
-});
-const performCalculation = {
-  '/': (first, second) => first / second,
-  '*': (first, second) => first * second,
-  '+': (first, second) => first + second,
-  '-': (first, second) => first - second,
-};
-
-const rappicargoOverlay = document.getElementById('rappicargo-overlay');
-const rappicargoModal = document.getElementById('rappicargo-modal');
-document.getElementById('cancel-rappicargo-btn').addEventListener('click', () => {
-  rappicargoModal.classList.remove('open');
-  rappicargoOverlay.classList.add('hidden');
-});
-document.getElementById('save-rappicargo-btn').addEventListener('click', () => {
-  const nameInput = document.getElementById('rappicargo-name-input');
-  const name = String(nameInput.value || '').trim();
-  if (name) {
-    addCode(name, 'RappiCargo', 'name');
-    nameInput.value = '';
-    rappicargoModal.classList.remove('open');
-    rappicargoOverlay.classList.add('hidden');
-  } else {
-    showNotification("Por favor, ingresa un nombre.", "error");
-  }
-});
-
 registerAuthHandlers();
 registerNavigationEvents({ setupRoleView, backToMenu, handleLogout });
+registerSettingsHandlers({ backToMenu });
+initCalculator();
+initRappicargo(addCode);
